+++ db/migrations/sqls/20220427152150-ddc-metric-down.sql
@@ -1,5 +1,5 @@
-DROP TABLE IF EXISTS ddc_metric;
+DROP VIEW IF EXISTS ddc_metric_view;
 
 DROP TABLE IF EXISTS ddc_metric_value;
 
-DROP VIEW IF EXISTS ddc_metric_view;
+DROP TABLE IF EXISTS ddc_metric;