+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## v0.11.0
 
 - Database migrations
+- Added `update-transactions-fees-cast-type` migration (compatible with the previous code version)
 
 ## v0.10.0