+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## vNext
 
 - Updated QAnet/Testnet urls
-- DDC metrics
+- DDC metrics (includes ddc-metric migration that is compatible with the code previous version)
 
 ## v0.11.0